Split tools out to a lib file
